% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{create_gemeente_url}
\alias{create_gemeente_url}
\title{Create valid url for 'gemeente'.}
\usage{
create_gemeente_url(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
valid url for 'gemeente' (character)
}
\description{
Create valid url for 'gemeente'.
}
