% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{create_house_htmls}
\alias{create_house_htmls}
\title{create a list of htmls with information on houses for sale.}
\usage{
create_house_htmls(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
list of htmls with information on houses for sale.
examples
x <- create_house_htmls("Doesburg" )
}
\description{
create a list of htmls with information on houses for sale.
}
