% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.adj_gemeente_str}
\alias{.adj_gemeente_str}
\title{Replace spaces and leave out characters 'Gemeente' in names of 'gemeente'
in order to create a valid url.}
\usage{
.adj_gemeente_str(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
Name of 'gemeente' in correct format to use in an url (character)
}
\description{
Replace spaces and leave out characters 'Gemeente' in names of 'gemeente'
in order to create a valid url.
}
