% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.create_gemeente_htmls}
\alias{.create_gemeente_htmls}
\title{Create list of html documents of all web pages of 'gemeente'.}
\usage{
.create_gemeente_htmls(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
valid html documents of 'gemeente' (character vector)
}
\description{
Create list of html documents of all web pages of 'gemeente'.
}
