% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.create_gemeente_urls}
\alias{.create_gemeente_urls}
\title{Create valid urls to all web pages of 'gemeente'.}
\usage{
.create_gemeente_urls(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
valid urls for 'gemeente' (character vector)
}
\description{
Create valid urls to all web pages of 'gemeente'.
}
