% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.create_house_urls}
\alias{.create_house_urls}
\title{Extract url's of houses for sale in 'gemeente'.}
\usage{
.create_house_urls(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
url's of houses for sale in 'gemeente' (character vector)
examples
.create_house_urls("Doesburg" )
}
\description{
Extract url's of houses for sale in 'gemeente'.
}
