% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.extract_address}
\alias{.extract_address}
\title{Extract address of house in html document.}
\usage{
.extract_address(html_doc)
}
\arguments{
\item{html_doc}{a html document (as read with function '.create_gemeente_htmls()')}
}
\value{
Address of house (character)
house_htmls <- fd_create_house_htmls("Doesburg" )
html_doc <- house_htmls[\link{1}]
.extract_address(html_doc)
}
\description{
Extract address of house in html document.
}
