% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.extract_core_urls_from_html}
\alias{.extract_core_urls_from_html}
\title{Use html document to extract the core of urls of houses for sale.}
\usage{
.extract_core_urls_from_html(html_doc)
}
\arguments{
\item{html_doc}{a html document (as read with function '.create_gemeente_htmls()')}
}
\value{
All core urls to houses for sale from html document (character vector)
}
\description{
Use html document to extract the core of urls of houses for sale.
}
