% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.is_house_info_link}
\alias{.is_house_info_link}
\title{TRUE if html attr (as read with html_attr(.,"href")) refers to a house for sale.}
\usage{
.is_house_info_link(html_attr)
}
\arguments{
\item{html_attr}{html attr (as read with html_attr(.,"href")) (character)}
}
\value{
TRUE if html attr (as read with html_attr(.,"href")) refers to a house for sale (boolean)
}
\description{
TRUE if html attr (as read with html_attr(.,"href")) refers to a house for sale.
}
