% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.is_new_construction}
\alias{.is_new_construction}
\title{TRUE if new construction}
\usage{
.is_new_construction(html_doc)
}
\arguments{
\item{html_doc}{a html document (as read with function '.create_gemeente_htmls()')}
}
\value{
TRUE if new construction (boolean)
}
\description{
TRUE if new construction
}
