% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.nr_of_pages}
\alias{.nr_of_pages}
\title{Return number of web pages associated with url of 'gemeente'}
\usage{
.nr_of_pages(url)
}
\arguments{
\item{url}{url of webpage (character)}
}
\value{
number of web pages associated with url of 'gemeente' (integer)
}
\description{
Return number of web pages associated with url of 'gemeente'
}
