% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{.plot_area}
\alias{.plot_area}
\title{Plot area of house.}
\usage{
.plot_area(html_doc)
}
\arguments{
\item{html_doc}{a html document (as read with function '.create_gemeente_htmls()')}
}
\value{
Plot area (numeric)
}
\description{
Plot area of house.
}
