% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_create_house_htmls}
\alias{fd_create_house_htmls}
\title{Create a list of htmls with information on houses for sale in 'gemeenten'.}
\usage{
fd_create_house_htmls(gemeenten)
}
\arguments{
\item{gemeenten}{Names of 'gemeente' (character vector)}
}
\value{
list with htmls with information on houses for sale in 'gemeenten'.
If multiple names are specified, a list of lists is returned (one list per 'gemeente').
examples
house_htmls <- fd_create_house_htmls(c("Doesburg","Zevenaar"))
}
\description{
Create a list of htmls with information on houses for sale in 'gemeenten'.
}
