% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_descr_match}
\alias{fd_descr_match}
\title{Check if pattern matches description of house in a list of html documents.}
\usage{
fd_descr_match(house_htmls, pattern)
}
\arguments{
\item{house_htmls}{list of html documents as read with function create_house_htmls().}

\item{pattern}{character string containing a regular expression.}
}
\value{
TRUE if matched; FALSE otherwise (boolean vector)
house_htmls <- fd_create_house_htmls("Doesburg" )
pattern <- "puntstuk|bron|waterbron|beregeningsbron|put|waterput|welput"
x <- fd_descr_match(house_htmls, pattern)
}
\description{
Check if pattern matches description of house in a list of html documents.
}
