% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_description}
\alias{fd_description}
\title{Extract 'description' of houses from a list of html documents.}
\usage{
fd_description(house_htmls)
}
\arguments{
\item{house_htmls}{list of html documents as read with function create_house_htmls().}
}
\value{
Description of houses (character vector)
examples
house_htmls <- fd_create_house_htmls("Doesburg" )
x <- description(house_htmls)
}
\description{
Extract 'description' of houses from a list of html documents.
}
