% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_gemeenten}
\alias{fd_gemeenten}
\title{Extract 'gemeenten' in list of html documents.}
\usage{
fd_gemeenten(house_htmls)
}
\arguments{
\item{house_htmls}{list of html documents as read with function create_house_htmls().}
}
\value{
'Gemeenten' (character vector)
house_htmls <- fd_create_house_htmls("Doesburg" )
x <- fd_plot_area(house_htmls)
}
\description{
Extract 'gemeenten' in list of html documents.
}
