% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_is_new_construction}
\alias{fd_is_new_construction}
\title{TRUE if new construction}
\usage{
fd_is_new_construction(house_htmls)
}
\arguments{
\item{house_htmls}{list of html documents as read with function create_house_htmls().}
}
\value{
TRUE if new construction (boolean vector)
examples
house_htmls <- fd_create_house_htmls("Doesburg" )
x <- fd_is_new_construction(house_htmls)
}
\description{
TRUE if new construction
}
