% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_plot_area}
\alias{fd_plot_area}
\title{Plot area of houses.}
\usage{
fd_plot_area(house_htmls)
}
\arguments{
\item{house_htmls}{list of html documents as read with function create_house_htmls().}
}
\value{
Plot area of houses (numeric vector)
examples
house_htmls <- fd_create_house_htmls("Doesburg" )
x <- fd_plot_area(house_htmls)
}
\description{
Plot area of houses.
}
