% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_type}
\alias{fd_type}
\title{Extract type of construction from a list of html documents.}
\usage{
fd_type(house_htmls)
}
\arguments{
\item{house_htmls}{list of html documents as read with function create_house_htmls().}
}
\value{
Type of construction (character vector)
examples
house_htmls <- fd_create_house_htmls("Doesburg" )
x <- fd_type(house_htmls)
}
\description{
Extract type of construction from a list of html documents.
}
