% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funda.R
\docType{package}
\name{funda}
\alias{funda}
\title{funda: Functions to explore the website 'funda.nl' on houses.}
\description{
This package exports the following functions:
}
\details{
\itemize{
\item \code{\link{create_house_htmls}}
}
}
