% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_utilities.R
\name{fd_create_house_htmls}
\alias{fd_create_house_htmls}
\title{Create a list of htmls with information on houses for sale in 'gemeente'.}
\usage{
fd_create_house_htmls(gemeente)
}
\arguments{
\item{gemeente}{Name of 'gemeente' (character)}
}
\value{
list of htmls with information on houses for sale in 'gemeente'.
examples
house_htmls <- fd_create_house_htmls("Doesburg" )
}
\description{
Create a list of htmls with information on houses for sale in 'gemeente'.
}
