% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funda.R
\docType{package}
\name{funda}
\alias{funda}
\title{funda: Functions to explore the website 'funda.nl'.}
\description{
This package exports the following functions:
}
\details{
\itemize{
\item \code{\link{fd_create_house_htmls}}
\item \code{\link{fd_addresses}}
\item \code{\link{fd_type}}
\item \code{\link{fd_plot_area}}
\item \code{\link{fd_gemeenten}}
\item \code{\link{fd_description}}
\item \code{\link{fd_descr_match}}
}
}
